docker/kyros/requirements.txt
Flask
requests
gunicorn
PyYAML